feat: 검색결과 비어있을 때 및 검색결과 작품 개수 명시

+++ app/src/main/java/com/teamwss/websoso/ui/normalExplore/adapter/NormalExploreItemType.kt
@@ -4,7 +4,7 @@ import com.teamwss.websoso.ui.normalExplore.model.NormalExploreModel.NovelModel
 
 sealed class NormalExploreItemType {
 
-    data class Header(val novelCount: Int) : NormalExploreItemType()
+    data class Header(val novelCount: Long) : NormalExploreItemType()
 
     data class Result(val novel: NovelModel) : NormalExploreItemType()
